% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransfer.R
\name{Authorize}
\alias{Authorize}
\title{Authorize}
\usage{
Authorize(app.id, app.key)
}
\arguments{
\item{app.id}{the app id}

\item{app.key}{the app key}
}
\value{
Nothing
}
\description{
Sets and stores the environment variables for the data transfer api
}
\examples{
\dontrun{ Authorize('yourappid', 'yourappkey') }
}
