% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransfer.R
\name{DataTransfer}
\alias{DataTransfer}
\title{DataTransfer}
\usage{
DataTransfer(product, startDate = NULL, endDate = NULL, directory = NULL)
}
\arguments{
\item{product}{the product to download}

\item{startDate}{the start date of the product to download}

\item{endDate}{the end date of the product to download}

\item{directory}{the directory you wish to download the datasets to. will only download new files}
}
\value{
a list of the file names downloaded
}
\description{
Given a product, downloads the datasets available for that product.
}
\examples{
\dontrun{files <- DataTransfer('skywise-conus-surface-analysis', directory = 'data')}
}
